% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bead_snake.R
\name{add_control_points}
\alias{add_control_points}
\title{Add control points to the middle of a two-point data frame}
\usage{
add_control_points(df, dir = "right", adjust = 3)
}
\arguments{
\item{df}{tibble subset of bead dataframe}
}
\description{
Add control points to the middle of a two-point data frame
}
