% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snake.R
\name{clean_districts}
\alias{clean_districts}
\title{Clean up district labels}
\usage{
clean_districts(label, ev)
}
\arguments{
\item{label}{text labels}

\item{ev}{electoral votes}
}
\description{
Replace state and district names with abbreviations (for districts with fewer 
EVs than letters)
}
