% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bead_snake.R
\name{compute_control_point}
\alias{compute_control_point}
\title{This function computes the control point based on weights passed in - 
the control point is (xwt, ywt) in relative proportion between the two rows 
in the data frame}
\usage{
compute_control_point(dfs, xwt = 0.5, ywt = 0.5, dir = 0, ...)
}
\description{
This function computes the control point based on weights passed in - 
the control point is (xwt, ywt) in relative proportion between the two rows 
in the data frame
}
