% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bead_snake.R
\name{compute_dir}
\alias{compute_dir}
\title{Compute the direction/y offset of the bezier control point}
\usage{
compute_dir(xx, adjust = 3)
}
\description{
Compute the direction/y offset of the bezier control point
}
