% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bead_snake.R
\name{cum_half_sum}
\alias{cum_half_sum}
\title{This function is useful for calculating the location of the center of the 
beads in various situations}
\usage{
cum_half_sum(x)
}
\description{
This function is useful for calculating the location of the center of the 
beads in various situations
}
