% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{elections}
\alias{elections}
\title{State-level election results of US presidential elections}
\format{A tibble with 564 rows and 10 columns
\describe{
  \item{year}{year of election}
  \item{state}{name of the state}
  \item{state_po}{two-letter abbreviation of state}
  \item{votes_dem}{number of votes for the Democratic candidate}
  \item{votes_rep}{number of votes for the Republican candidate}
  \item{perc_dem}{percent of votes for the Democratic candidate}
  \item{perc_rep}{percent of votes for the Republican candidate}
  \item{totalvotes}{number of total votes}
  \item{cand_dem}{name of the Democratic candidate}
  \item{cand_rep}{name of the Republican candidate}
}}
\usage{
elections
}
\description{
Wide format of the `USelection` data, watered down to democratic and republican votes.
}
\keyword{datasets}
