% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electoral_building.R
\name{electoral_building}
\alias{electoral_building}
\title{Electoral Building}
\usage{
electoral_building(state_district, electoral_votes, perc_dem, perc_rep)
}
\arguments{
\item{state_district}{state/district name}

\item{electoral_votes}{number of electoral votes}

\item{perc_dem}{percent democrats}

\item{perc_rep}{percent republicans}
}
\description{
Electoral Building
}
