% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snake.R
\name{ggsnake}
\alias{ggsnake}
\title{Create an Electoral Snake Plot}
\usage{
ggsnake(.data)
}
\arguments{
\item{.data}{data frame}

\item{pct_diff}{percent difference between republican and democrat votes/polls. Democrat advantage should be a negative number.}

\item{electoral_votes}{number of electoral votes in each district}

\item{district}{}
}
\description{
Create an Electoral Snake Plot
}
