% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_combine.R
\name{polls_filter}
\alias{polls_filter}
\title{Select polls}
\usage{
polls_filter(polls, method = "last_k", k = 5, since = today() - weeks(4))
}
\arguments{
\item{polls}{data frame of polling results as returned by `fivethirtyeight_update``}

\item{method}{which method do we use to include polls: "last_k" includes the last `k` polls, "since" includes polls as specified in `since`.}

\item{k}{include k most recent polls}

\item{since}{date after which polls to include (based on `date` or `end_date`).}
}
\description{
Filter polls based on k most recent by state or since a set date.
}
