% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{presidential_polls}
\alias{presidential_polls}
\title{Polls for the presidential general election}
\format{A tibble with 5312 rows and 37 columns
\describe{
  \item{question_id}{numeric}
  \item{poll_id}{numeric}
  \item{cycle}{numeric}
  \item{state}{character}
  \item{pollster_id}{numeric}
  \item{pollster}{character}
  \item{sponsor_ids}{numeric}
  \item{sponsors}{character}
  \item{display_name}{character}
  \item{pollster_rating_id}{numeric}
  \item{pollster_rating_name}{character}
  \item{fte_grade}{character}
  \item{sample_size}{numeric}
  \item{population}{character}
  \item{population_full}{character}
  \item{methodology}{character}
  \item{office_type}{character}
  \item{seat_number}{numeric}
  \item{seat_name}{logical}
  \item{start_date}{Date}
  \item{end_date}{Date}
  \item{election_date}{Date}
  \item{sponsor_candidate}{character}
  \item{internal}{logical}
  \item{partisan}{character}
  \item{tracking}{logical}
  \item{nationwide_batch}{logical}
  \item{ranked_choice_reallocated}{logical}
  \item{created_at}{character}
  \item{notes}{character}
  \item{url}{character}
  \item{stage}{character}
  \item{race_id}{numeric}
  \item{answer}{character}
  \item{candidate_name}{character}
  \item{candidate_party}{character}
  \item{pct}{numeric}
}}
\source{
\url{https://projects.fivethirtyeight.com/polls-page/president_polls.csv}
}
\usage{
presidential_polls
}
\description{
Scraped from FiveThirtyEight's website of most recent polls on July 28 2020.
}
\examples{
library(ggplot2)
library(dplyr)

data(presidential_polls)
presidential_polls \%>\% 
  filter(candidate_party \%in\% c("DEM", "REP")) \%>\% 
  filter(!is.na(state)) \%>\%
  ggplot(aes(x = end_date, y = pct)) + 
    geom_point(aes(colour = candidate_party)) +
    facet_wrap(~state) +
    scale_colour_party()

# national polls
presidential_polls \%>\% 
  filter(candidate_party \%in\% c("DEM", "REP")) \%>\% 
  filter(answer \%in\% c("Biden", "Trump")) \%>\% 
  filter(is.na(state)) \%>\%
  ggplot(aes(x = end_date, y = pct, 
             colour = candidate_party)) + 
    geom_point(alpha = 0.3) +
    geom_smooth(se = FALSE, method="gam") +
    scale_colour_party() +
    theme(legend.position="bottom")
}
\keyword{datasets}
