% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polls.R
\name{rcp_update}
\alias{rcp_update}
\title{This function returns the latest polls from Real Clear Politics}
\usage{
rcp_update(type = "state_president")
}
\arguments{
\item{type}{one of "national_president", "state_president", 
"state_of_the_union". Deafults to "state_president"; 
senate, governor tables are not optimally tidy.}
}
\value{
a table of poll links, with results for each candidate
}
\description{
This function returns the latest polls from Real Clear Politics
}
\examples{
rcp_update()
rcp_update("national_president")
}
