% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scales.R
\name{scale_color_party}
\alias{scale_color_party}
\title{Color scale for two parties}
\usage{
scale_color_party(...)
}
\arguments{
\item{...}{passed on to scale_color_manual}
}
\description{
Color scale for two parties
}
