% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more_polls.R
\name{scale_color_partygrad}
\alias{scale_color_partygrad}
\title{Color scale for gradient between two parties}
\usage{
scale_color_partygrad(...)
}
\arguments{
\item{...}{passed on to scale_color_manual}
}
\description{
Color scale for gradient between two parties
}
