% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more_polls.R
\name{scale_colour_party}
\alias{scale_colour_party}
\title{Color scale for two parties}
\usage{
scale_colour_party(...)
}
\arguments{
\item{...}{passed on to scale_colour_manual}
}
\description{
Color scale for two parties
}
