% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scales.R
\name{scale_colour_party_binned}
\alias{scale_colour_party_binned}
\title{Color scale for binned gradient between two parties}
\usage{
scale_colour_party_binned(...)
}
\arguments{
\item{...}{passed on to scale_*_gradient2}
}
\description{
Color scale for binned gradient between two parties
}
