% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scales.R
\name{scale_colour_partygrad}
\alias{scale_colour_partygrad}
\title{Color scale for gradient between two parties}
\usage{
scale_colour_partygrad(...)
}
\arguments{
\item{...}{passed on to scale_*_gradient2}
}
\description{
Color scale for gradient between two parties
}
