% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more_polls.R
\name{scale_fill_party}
\alias{scale_fill_party}
\title{Fill scale for two parties}
\usage{
scale_fill_party(...)
}
\arguments{
\item{...}{passed on to scale_fill_manual}
}
\description{
Fill scale for two parties
}
