% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scales.R
\name{scale_fill_party_binned}
\alias{scale_fill_party_binned}
\title{Fill scale for binned gradient between two parties}
\usage{
scale_fill_party_binned(...)
}
\arguments{
\item{...}{passed on to scale_*_gradient2}
}
\description{
Fill scale for binned gradient between two parties
}
\examples{

library(dplyr)
electoral_votes_2016 \%>\%
mutate(diff = perc_rep - perc_dem, party = ifelse(perc_dem > perc_rep, "Dem", "Rep")) \%>\%
ggsnake(order = diff, fill = diff, label = state_district, color = party, size = electoral_votes) + 
scale_fill_party_binned() + scale_color_party()
}
