% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bead_snake.R
\name{set_up_beads}
\alias{set_up_beads}
\title{Helper function}
\usage{
set_up_beads(df, height = NULL, buffer = NULL)
}
\arguments{
\item{df}{data frame}

\item{height}{overall height}

\item{buffer}{space between beads}
}
\description{
Sets up beads based on initial information
This should eventually become stat_beads or something like that, I think.
}
