% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{snake_path}
\alias{snake_path}
\title{Electoral snake path data}
\format{A tibble with 3278 rows and 4 columns
\describe{
  \item{ev}{electoral vote number}
  \item{order}{path plotting order}
  \item{x}{x coordinate}
  \item{y}{y coordinate}
}}
\usage{
snake_path
}
\description{
A winding path of 538 approximately equal segments to be used for making
the electoral snake plot.
}
\keyword{datasets}
