% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bead_snake.R
\name{split_df_bezier}
\alias{split_df_bezier}
\title{Split the bead dataframe into directional segments}
\usage{
split_df_bezier(df, height = NULL, buffer = NULL)
}
\arguments{
\item{df}{tibble with bead information}

\item{height}{plot height (in sqrt(electoral votes/pi) units)}

\item{buffer}{space between circles}
}
\description{
Split the bead dataframe into directional segments
}
