% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{USelections}
\alias{USelections}
\title{Results of the US Presidential elections by state 1976-2016}
\format{An object of class data.frame with 3740 rows and 14 columns.
\describe{
\item{year}{Year of the election between 1976 and 2016}
\item{state}{Name of the US state}
\item{state_po}{U.S. postal code state abbreviation}
\item{state_fips}{State FIPS code}
\item{state_cen}{U.S. Census state code}
\item{state_ic}{ICPSR state code}
\item{office}{U.S. President}
\item{candidate}{Name of the candidate}
\item{party}{Party of the candidate}
\item{writein}{Vote totals associated with write-in candidates}
\item{candidatevotes}{Votes received by this candidate for this particular party}
\item{totalvotes}{Total number of votes cast for this election}
\item{version}{Date of final version}
\item{notes}{logical vector}
}}
\source{
\url{https://doi.org/10.7910/DVN/42MVDX}
}
\usage{
USelections
}
\description{
A dataset containing results of all US Presidential elections by state from 1976 to 2016.
This data was provided by the MIT Election Data + Science Lab.
}
\references{
MIT Election Data and Science Lab, 2017, "U.S. President 1976–2016", \url{https://doi.org/10.7910/DVN/42MVDX}, Harvard Dataverse, V5.
}
\keyword{datasets}
