% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ushex}
\alias{ushex}
\title{Hexbin map of the United States}
\format{An object of class data.frame with 357 rows and 8 columns.
\describe{
\item{long}{geographic longitude}
\item{lat}{geographic latitude}
\item{order}{order in which nodes are connected}
\item{group}{group}
\item{state}{name of the state}
\item{abbr}{US postal code state abbreviations}
\item{centerX}{center of the hexagon in longitude}
\item{centerY}{center of the hexagon in latitude}
}}
\source{
\url{https://team.carto.com/u/andrew/tables/andrew.us_states_hexgrid/public/map}
}
\usage{
ushex
}
\description{
Hexbin map of the United states modified to include hexagon centers and state abbreviations.
The original hexbin map is provided by Andrew X Hill at \url{https://team.carto.com/u/andrew/tables/andrew.us_states_hexgrid/public/map}.
}
\references{
Andrew X Hill
}
\keyword{datasets}
