% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bead_snake.R
\name{bead_snake_plot}
\alias{bead_snake_plot}
\title{Plot state electoral votes as beads on a string}
\usage{
bead_snake_plot(
  data,
  height = max(sqrt(electoral_votes_2016$electoral_votes) * 8),
  buffer = min(sqrt(electoral_votes_2016$electoral_votes) * 3)
)
}
\arguments{
\item{data}{data frame with format similar to electoral_votes_2016}

\item{height}{column height, default is carefully chosen based on number of electoral votes}

\item{buffer}{space between beads}
}
\description{
TODO: Make this work with NSE
}
\examples{

data("electoral_votes_2016")
bead_snake_plot(electoral_votes_2016)
bead_snake_plot(electoral_votes_2016, height = 30, buffer = 3)
}
