% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electoral_building.R
\name{electoral_building}
\alias{electoral_building}
\title{Electoral Building}
\usage{
electoral_building(state_district, electoral_votes, perc_dem, perc_rep, source)
}
\arguments{
\item{state_district}{state/district name}

\item{electoral_votes}{number of electoral votes}

\item{perc_dem}{percent democrats}

\item{perc_rep}{percent republicans}

\item{source}{polls or election results?}
}
\description{
Electoral Building
}
\examples{
library(ggplot2)
electoral_building(state_district = electoral_votes_2016$state_district, 
                   electoral_votes = electoral_votes_2016$electoral_votes, 
                   perc_dem = electoral_votes_2016$perc_dem, 
                   perc_rep = electoral_votes_2016$perc_rep,
                   source = "election"
                   ) +
  scale_color_party("Party") +
  scale_fill_party("Party") +
  theme(legend.position = "bottom")
}
