% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bead_snake.R
\name{mysumfun}
\alias{mysumfun}
\title{Helper function}
\usage{
mysumfun(x)
}
\arguments{
\item{x}{vector}
}
\description{
Summarize by taking the mean (if numeric), the unique value 
(if there's only one), and NA otherwise
}
