% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_combine.R
\name{polls_plus}
\alias{polls_plus}
\title{Average polls and supplement with election data}
\usage{
polls_plus(polls, election, k = 5)
}
\arguments{
\item{polls}{tibble result from `fivethirteight_update`}

\item{election}{tibble of election results (with number of electoral votes)}

\item{k}{include k most recent polls}
}
\value{
tibble of average polls supplemented by last k polls
polls_plus(presidential_polls %>% filter(population %in% c("lv", "rv")),
           electoral_votes_2016) %>% head()
}
\description{
Average polls and supplement with election data
}
