% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_scales.R
\name{scale_fill_partygrad}
\alias{scale_fill_partygrad}
\title{Fill scale for gradient between two parties}
\usage{
scale_fill_partygrad(...)
}
\arguments{
\item{...}{passed on to scale_*_gradient2}
}
\description{
Fill scale for gradient between two parties
}
