% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bead_snake.R
\name{set_up_beads}
\alias{set_up_beads}
\title{Sets up beads based on initial information
This should eventually become stat_beads or something like that, I think.}
\usage{
set_up_beads(df, height = NULL, buffer = NULL)
}
\description{
Sets up beads based on initial information
This should eventually become stat_beads or something like that, I think.
}
